Improve instructions test a bit.
- Print assumed test result. For the assumption to hold, check the
binary is actually expected. If the binary is not same, the test might
still run fine, but we don't know whether the address cpu stuck is
success or failure.
- Reduce elapsed time decimals.
- Let CPU.run return some kind of exit status.
- Test stuck jump after the PC has been changed. This way CPU.pc
contains the address where the jump exists and not the address after
that. Makes easier to check where the CPU stuck, and probably feels more
consistent compared to last instructions log.